refactor(deck): 로스터 빌더 공용 모듈로 추출 (#62)
