+++ src/features/deck-recommendation/actions/recommendDeck.ts
@@ -1,7 +1,6 @@
 'use server';
 
 import { createClient } from '@/shared/lib/supabase/server';
-import type { PokemonType } from '@/shared/types/pokemon';
 import { RecommendRequestSchema } from '../model/schemas';
 import type { RecommendedDeck, RecommendResponse, RosterPokemon } from '../model/schemas';
 import {
@@ -12,6 +11,7 @@ import {
   filterSpeed,
   filterCounter,
 } from '../lib/rule-engine';
+import { loadOwnedRoster } from '../lib/roster';
 import { generateRecommendation, RECOMMENDATION_MODEL } from '../lib/gemini';
 import { fallbackRecommendation } from '../lib/fallback';
 import {
@@ -21,79 +21,6 @@ import {
   setCachedRecommendation,
   checkRateLimit,
 } from '../lib/cache';
-import rawMovesJson from '../../../../data/moves.json';
-import rawPokemonMovesJson from '../../../../data/pokemon-moves.json';
-
-type MoveEntry = {
-  id: string;
-  koName: string;
-  type: string;
-  damageClass: string;
-  power: number;
-  accuracy: number;
-  pp: number;
-};
-
-const movesData = rawMovesJson as Record<string, MoveEntry>;
-const pokemonMovesData = rawPokemonMovesJson as Record<string, string[]>;
-
-async function loadRoster(
-  supabase: Awaited<ReturnType<typeof createClient>>,
-  userId: string,
-): Promise<RosterPokemon[]> {
-  const { data: trainerPokemons } = await supabase.from('owned_pokemon').select('dex_id').eq('user_id', userId);
-
-  if (!trainerPokemons || trainerPokemons.length === 0) return [];
-
-  // 같은 종을 여러 마리 보유할 수 있으므로 dex_id 중복 제거
-  const dexIds = [...new Set(trainerPokemons.map((p: { dex_id: number }) => p.dex_id))];
-
-  const { data: species } = await supabase
-    .from('pokemon_species')
-    .select('dex_id, ko_name, artwork_url, type1_id, type2_id, base_hp, base_atk, base_def, base_spd')
-    .in('dex_id', dexIds);
-
-  if (!species) return [];
-
-  return species.map(
-    (s: {
-      dex_id: number;
-      ko_name: string;
-      artwork_url: string | null;
-      type1_id: PokemonType;
-      type2_id: PokemonType | null;
-      base_hp: number;
-      base_atk: number;
-      base_def: number;
-      base_spd: number;
-    }) => {
-      const moveIds = pokemonMovesData[String(s.dex_id)] ?? [];
-      const moves = moveIds.slice(0, 4).map((id) => {
-        const m = movesData[id];
-        return {
-          id,
-          koName: m?.koName ?? id,
-          type: (m?.type ?? 'normal') as PokemonType,
-          power: m && m.power > 0 ? m.power : null,
-          statusEffect: m?.damageClass === 'status' ? 'status' : null,
-        };
-      });
-
-      return {
-        dexId: s.dex_id,
-        koName: s.ko_name,
-        artworkUrl: s.artwork_url ?? '',
-        type1: s.type1_id,
-        type2: s.type2_id,
-        level: 1,
-        baseAtk: s.base_atk,
-        baseSpd: s.base_spd,
-        baseStatTotal: s.base_hp + s.base_atk + s.base_def + s.base_spd,
-        moves,
-      };
-    },
-  );
-}
 
 function trimDesc(deck: RecommendedDeck): RecommendedDeck {
   return deck.description.length <= 14 ? deck : { ...deck, description: deck.description.slice(0, 14) };
@@ -117,7 +44,7 @@ export async function recommendDeck(rawInput: unknown): Promise<RecommendRespons
     return { ok: false, message: '로그인이 필요합니다' };
   }
 
-  const roster = await loadRoster(supabase, user.id);
+  const roster = await loadOwnedRoster(supabase, user.id);
 
   if (roster.length < 3) {
     return { ok: false, message: '추천을 받으려면 최소 3마리의 포켓몬이 필요합니다' };
@@ -223,7 +150,7 @@ export async function recommendHomeDecks(): Promise<HomeDecksResponse> {
   const err: RecommendResponse = { ok: false, message: '로그인이 필요합니다' };
   if (!user) return { decks: [err, err] };
 
-  const roster = await loadRoster(supabase, user.id);
+  const roster = await loadOwnedRoster(supabase, user.id);
 
   if (roster.length < 3) {
     const e: RecommendResponse = { ok: false, message: '추천을 받으려면 최소 3마리의 포켓몬이 필요합니다' };
